+++ lib/flutter_folderview.dart
@@ -11,4 +11,5 @@ export 'themes/folder_view_scrollbar_theme.dart';
 export 'themes/folder_view_spacing_theme.dart';
 export 'themes/folder_view_theme.dart';
 export 'themes/parent_node_theme.dart';
+export 'themes/node_tooltip_theme.dart';
 export 'widgets/folder_view.dart';